requirements.txt
tensorflow-gpu==1.12
tensorflow-hub==0.7.0
h5py
nltk
pyhocon
scipy
sklearn
